Updates package name to org.gauge.gradle and new extension
