+++ .github/workflows/build.yml
@@ -4,23 +4,26 @@ on: [push, pull_request]
 
 jobs:
   build:
-
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
         os: [ubuntu-latest, windows-latest]
-
     steps:
-    - uses: actions/checkout@v1
-    - name: Set up JDK
-      uses: actions/setup-java@v1
+    - uses: actions/checkout@v4
+    - name: Setup java
+      uses: actions/setup-java@v3
       with:
-        java-version: '12.x.x'
-
+        distribution: 'temurin'
+        java-version: '11'
+    - uses: getgauge/setup-gauge@master
+      with:
+        gauge-version: '1.5.4'
+        gauge-plugins: java, html-report
     - name: Build with Gradle on ubuntu
       if: matrix.os != 'windows-latest'
-      run: ./gradlew clean build
-
+      run: |
+        ./gradlew clean build gauge
     - name: Build with Gradle on windows
       if: matrix.os == 'windows-latest'
-      run: ./gradlew.bat clean build
+      shell: pwsh
+      run: .\gradlew.bat clean build gauge
