+++ .github/workflows/release.yml
@@ -12,16 +12,15 @@ jobs:
       GRADLE_PUBLISHKEY: '${{ secrets.GRADLE_PUBLISHKEY }}'
       GRADLE_SECRET: '${{ secrets.GRADLE_SECRET }}'
     steps:
-      - uses: actions/checkout@v1
-      - name: Set up JDK
-        uses: actions/setup-java@v1
+      - uses: actions/checkout@v4
+      - name: Setup java
+        uses: actions/setup-java@v3
         with:
-          java-version: '12.x.x'
-
+          distribution: 'temurin'
+          java-version: '11'
       - name: Build artifacts
         run: |
           ./gradlew clean build
-
       - name: Upload to gradle portal
         run: |
           ./gradlew publishPlugins -Pgradle.publish.key=$GRADLE_PUBLISHKEY -Pgradle.publish.secret=$GRADLE_SECRET
