+++ build.gradle
@@ -1,11 +1,6 @@
-allprojects {
+subprojects {
     repositories {
         mavenLocal()
         mavenCentral()
     }
 }
-
-Properties plugin = new Properties()
-plugin.load(project.file('./plugin.properties').newDataInputStream())
-group plugin.getProperty("groupId")
-version = plugin.getProperty("version")
