+++ gradle.properties
@@ -1,0 +1,1 @@
+version=2.0.0
