--- plugin.properties
@@ -1,12 +1,0 @@
-siteUrl=https://github.com/getgauge/gauge-gradle-plugin
-gitUrl=https://github.com/getgauge/gauge-gradle-plugin.git
-
-version=2.0.0
-name=gauge-gradle-plugin
-displayName=Gauge
-description=Gradle plugin for Gauge, the open source test automation tool developed by ThoughtWorks.
-versionDescription=Remove specsDir if --failed or --repeat flags are given
-
-groupId=org.gauge.gradle
-pluginKey=org.gauge
-artifactId=gauge-gradle-plugin
