+++ plugin/build.gradle
@@ -1,12 +1,9 @@
 plugins {
-    id 'java'
+    id 'java-gradle-plugin'
     id 'com.gradle.plugin-publish' version '1.2.1'
 }
 
-Properties plugin = new Properties()
-plugin.load(project.file('../plugin.properties').newDataInputStream())
-group plugin.getProperty("groupId")
-version = plugin.getProperty("version")
+group = 'org.gauge.gradle'
 
 java {
     sourceCompatibility = JavaVersion.VERSION_11
@@ -33,10 +30,10 @@ gradlePlugin {
         website = "https://github.com/getgauge/gauge-gradle-plugin"
         gauge {
             id = "org.gauge"
-            implementationClass = "com.thoughtworks.gauge.gradle.GaugePlugin"
+            implementationClass = "org.gauge.gradle.GaugePlugin"
             displayName = "Gauge"
             description = "Gradle plugin for Gauge, the open source test automation tool developed by ThoughtWorks."
-            tags.addAll('gauge', 'thoughtworks')
+            tags.addAll('gauge', 'thoughtworks', 'testing')
         }
     }
 }
