+++ plugin/src/main/java/org/gauge/gradle/GaugeConstants.java
@@ -1,0 +1,14 @@
+package org.gauge.gradle;
+
+class GaugeConstants {
+    public static final String GAUGE_PLUGIN_ID = "org.gauge";
+    public static final String GAUGE_EXTENSION_ID = "gauge";
+    public static final String ENVIRONMENT = "env";
+    public static final String TAGS = "tags";
+    public static final String SPECS_DIR = "specsDir";
+    public static final String IN_PARALLEL = "inParallel";
+    public static final String NODES = "nodes";
+    public static final String ADDITIONAL_FLAGS = "additionalFlags";
+    public static final String GAUGE_ROOT = "gaugeRoot";
+
+}
