--- sample/src/spec/java/com/thoughtworks/gauge/gradle/sample/Test.java
@@ -1,13 +1,0 @@
-package com.thoughtworks.gauge.gradle.sample;
-
-public class Test {
-    private String value;
-
-    public Test(String value) {
-        this.value = value;
-    }
-
-    public String getValue() {
-        return value;
-    }
-}
