+++ sample/src/spec/java/org/gauge/gradle/sample/Main.java
@@ -1,0 +1,8 @@
+package org.gauge.gradle.sample;
+
+public class Main {
+    public static void main(String[] args) {
+        Test test = new Test("test_value");
+        System.out.println("org.gauge.gradle.sample.Test Value is : " + test.getValue());
+    }
+}
