+++ sample/src/spec/java/org/gauge/gradle/sample/Test.java
@@ -1,0 +1,13 @@
+package org.gauge.gradle.sample;
+
+public class Test {
+    private final String value;
+
+    public Test(String value) {
+        this.value = value;
+    }
+
+    public String getValue() {
+        return value;
+    }
+}
